+++ terrain/voxel_mesh_block.h
@@ -36,14 +36,14 @@ public:
 	// Visuals
 
 	void set_mesh(
-			Ref<Mesh> mesh, DirectMeshInstance::GIMode gi_mode, RenderingServer::ShadowCastingSetting shadow_casting);
+			Ref<Mesh> mesh, GeometryInstance3D::GIMode gi_mode, RenderingServer::ShadowCastingSetting shadow_casting);
 	Ref<Mesh> get_mesh() const;
 	bool has_mesh() const;
 	void drop_mesh();
 
 	// Note, GIMode is not stored per block, it is a shared option so we provide it in several functions.
 	// Call this function only if the mesh block already exists and has not changed mesh
-	void set_gi_mode(DirectMeshInstance::GIMode mode);
+	void set_gi_mode(GeometryInstance3D::GIMode mode);
 
 	// Note, ShadowCastingSetting is not stored per block, it is a shared option so we provide it in several functions.
 	// Call this function only if the mesh block already exists and has not changed mesh