+++ doc/source/changelog.md
@@ -50,6 +50,7 @@ Semver is not yet in place, so each version can have breaking changes, although 
         - Added `get_all_item_ids()` to allow iterating over all items of a library
     - `VoxelLibraryMultiMeshItem `:
         - Added `render_layer` property (thanks to m4nu3lf)
+        - Added `gi_mode` property
         - Node groups on the template scene are now added to instance colliders if present
     - `VoxelLodTerrain`:
         - Added debug drawing for modifier bounds
@@ -96,6 +97,8 @@ Semver is not yet in place, so each version can have breaking changes, although 
         - Changed the list of models to be handled by a typed array instead of individual properties. When opened in the editor, old resources will get converted. Re-save them to make the conversion persist.
     - `VoxelBlockyModel`:
         - The class was split into several subclasses for each type of geometry. When opened in the editor, old resources will get converted, but only if they are part of a `VoxelBlockyLibaray`. They won't work if they are individual resource files.
+    - `VoxelNode`:
+        - Removed `GIMode` enum, replaced with `GeometryInstance3D.GIMode`
 
 
 1.0 - 12/03/2023 - `godot4.0`